% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cS2G_link.R
\name{cs2g_link}
\alias{cs2g_link}
\title{Combined S2G Strategy}
\usage{
cs2g_link(
  pheno_name,
  SNP_data,
  s2g_folder,
  save_folder,
  keep_intermediate_file = FALSE,
  intermediate_file_folder = NULL,
  gene_remove_list = NULL,
  other_weights = FALSE,
  other_precisions = FALSE,
  causal_SNP = TRUE,
  assumed_PIP = NULL,
  gene_annot_file = NULL,
  keep_unmatched_gene = FALSE
)
}
\arguments{
\item{pheno_name}{The pheno or trait include in your study, for example "height"; If you have two or more pheno, please separate them with commas, for example "height,BMI".}

\item{SNP_data}{The path point to your causal SNP or lead SNP (it is not recommended yet) result, but you should never use them together. The data should have each line representing a SNP, and it should contain at least CHR, SNP, BP, PHENO (if multiple phenotypes are provided in pheno_name), and PIPS (if they are causal SNPs).}

\item{s2g_folder}{The path point to 'S2G_bed' mapping file. This file should be compressed and contain "Exon", "Promoter", "finemappedciseQTLs_eQTLGen", finemappedciseQTLs_GTeX", "EpiMap", "ABC", "Ciceroblood" seven subfolders. The specific files within these subfolders should be uncompressed}

\item{save_folder}{The path where you want to store the result file.}

\item{keep_intermediate_file}{Whether you want to store intermediate files, default to False.}

\item{intermediate_file_folder}{If you want to store intermediate files, you should provide a specific storage path. It's default to NULL}

\item{gene_remove_list}{We suggest deleting non coding genes to clarify the results, as we have included a list of genes that should be removed. If you do not want to exclude, please select False; If you wish to add an additional list, please provide the specific path of the single column text without a header. It's default to NULL.}

\item{other_weights}{We used the optimal weights provided in the Gazal manuscript as the integrated weights for the seven strategies. If you have any other weights you would like to use, please provide the specific file. This data file should not have a header. The first column corresponds to the names of the seven strategies used (Please refer to the description of the s2d_folder parameter above for details), the second column should correspond to specific weights. That is to say, this file should have seven rows and two columns. It's default to NULL.}

\item{other_precisions}{We used the precision of these seven strategies from the Gazal manuscript in the validation set as part of the confidence score calculation. If you have any other data you would like to use, please enter it. The specific requirements of the document are the same as' other weights'. It's default to NULL.}

\item{causal_SNP}{Whether the SNPs included in 'SNP_data' are causal SNPs, default to TRUE.}

\item{assumed_PIP}{If your SNP does not come from fine mapping, you need to provide a hypothetical posterior probability to calculate the confidence score. We recommend using a conservative data (default is 0.5), although this may require support from the other two types of evidence for all linking results, it can effectively reduce false positive results.}

\item{gene_annot_file}{In order to be compatible with the other two types of strategies, we need to convert the gene names obtained from s2g into ENSGID. We have provided the corresponding data, but if you want to use another list, please specify it. It's default to NULL.}

\item{keep_unmatched_gene}{Whether to retain genes that failed to convert ENSGID, default to False}
}
\value{
A data.frame named pheno_cS2G.csv
}
\description{
Use locus-based gene linking by combined SNP-to-GENE (cS2G) strategy
}
