% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtl_data_preparation.R
\name{eqtl_data_preparation}
\alias{eqtl_data_preparation}
\title{Prepare eQTL summary statistics for colocalization}
\usage{
eqtl_data_preparation(
  eqtl_data_folder,
  eqtl_data_tissue = NULL,
  save_folder = NULL,
  smr_exe,
  extract_gene_list,
  smr_file_delete = TRUE,
  complete_file_delete = TRUE,
  maf_impute = TRUE,
  maf_impute_file = NULL
)
}
\arguments{
\item{eqtl_data_folder}{Directory containing the raw eQTL data.}

\item{eqtl_data_tissue}{Character vector. A vector of specific tissue names to process (e.g., \code{c("Whole_Blood", "Liver")}). If \code{NULL} (default), the function detects and processes all available tissues in the folder. Note: Input names are validated against the standard GTEx V8/eQTLGen tissue list.}

\item{save_folder}{Directory to save output. Defaults to eqtl_data_folder.}

\item{smr_exe}{Path to the SMR executable file.}

\item{extract_gene_list}{Path to a file containing a list of genes to extract.}

\item{smr_file_delete}{Logical. Whether to delete intermediate SMR binary files after conversion.}

\item{complete_file_delete}{Logical. Whether to delete the huge converted text file after splitting.}

\item{maf_impute}{Logical. Whether to impute MAF from an external reference.}

\item{maf_impute_file}{Path to the external MAF reference file (containing 'SNP' and 'MAF' columns). By default, the function uses the GTEx V8 MAF reference data provided internally by this package.}
}
\description{
Due to its high accessibility and standardized formatting, this function is specifically designed to process GTEx v8 (accessible at \url{https://yanglab.westlake.edu.cn/software/smr/#DataResource}) and eQTLGEN (accessible at in \url{https://www.eqtlgen.org/cis-eqtls.html}) cis-eQTL summary statistics in \strong{SMR binary format}.

\strong{Important Notes:}
\itemize{
\item \strong{Data Completeness:} Do \strong{not} use "Lite" summary statistics (which only contain significant SNPs reaching a certain threshold). Colocalization analysis requires \strong{full summary statistics} (including non-significant SNPs) to accurately evaluate the posterior probability of a shared causal variant.
\item \strong{Other Data Sources:} While optimized for YangLab's format, you may use GTEx, eQTLGen, or other QTL summary statistics from different sources. However, if bypassing this function's processing, you must manually format your data into individual text files per gene, matching the output structure of this function. Matching relies on names here, so the downloaded summary file should be in compressed format and the name should not be changed.
\item \strong{MAF Imputation:} The SMR binary files provided by YangLab lack Minor Allele Frequency (MAF) information. This function automatically imputes MAF using the \strong{GTEx V8 reference data} bundled with this package. (Note: eQTLGen data does not require imputation).
}
}
