% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nominated_gene_extract.R
\name{nominated_gene_extract}
\alias{nominated_gene_extract}
\title{Extract genes for colocalization-based linking (i.e., nominated genes)}
\usage{
nominated_gene_extract(
  pheno_name,
  cs2g_result_folder,
  pops_result_folder,
  save_folder
)
}
\arguments{
\item{save_folder}{}
}
\description{
Use this function to organize the list of potentially relevant genes obtained from locus-based linking (cS2G) and similarity-based linking (PoPs)—referred to as “nominated genes”—for subsequent colocalization-based linking work.
}
