% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pops_result_process.R
\name{pops_result_processing}
\alias{pops_result_processing}
\title{Clean the PoPs result}
\usage{
pops_result_processing(
  pheno_name,
  SNP_data,
  pops_result_folder,
  pops_result_suffix = NULL,
  save_folder,
  annot_file = NULL,
  filter_gene = TRUE,
  replace_list = NULL
)
}
\arguments{
\item{pheno_name}{Character vector. Phenotypes to process.}

\item{SNP_data}{String (file path) or Data Frame. Must contain CHR, SNP, BP.}

\item{pops_result_folder}{String. Directory containing raw PoPS results.}

\item{pops_result_suffix}{String. Default ".preds".}

\item{save_folder}{String. Output directory.}

\item{annot_file}{String (path) or Data Frame. Gene annotation (CHR, START, END, GENE).}

\item{filter_gene}{We suggest deleting non coding genes to clarify the results. If you do not want to exclude, please select False. It's default to TRUE.}

\item{replace_list}{String (path) or Data Frame. Old ID -> New ID dictionary.}
}
\description{
Use this function to convert gene names or delete specific genes (such as non-protein coding genes).
}
