% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_process.R
\name{result_process}
\alias{result_process}
\title{Process and Summarize mdS2G Results}
\usage{
result_process(
  coloc_result,
  SNP_data_create_above,
  cs2g_cred_data,
  save_folder
)
}
\arguments{
\item{coloc_result}{Character string. File path to the colocalization results.
This specifically refers to the \strong{"coloc.csv"} file generated by the \code{eqtl_coloc} function.}

\item{SNP_data_create_above}{Character string. File path to the SNP mapping data.
This specifically refers to the \strong{"SNP_data_for_coloc.csv"} file generated by the \code{nominated_gene_extract} function.}

\item{cs2g_cred_data}{Character string. File path to the high-confidence cS2G results.
This specifically refers to the \strong{"cs2g_cred_gene.csv"} file generated by the \code{nominated_gene_extract} function.}

\item{save_folder}{Character string. Directory path to save the processed outputs.}
}
\value{
Outputs several CSV and TXT files to \code{save_folder}.
}
\description{
This function integrates results from colocalization (Coloc), locus-based linking (cS2G), and similarity-based linking (PoPS).
It identifies "Credible Genes" supported by multiple lines of evidence (e.g., PoPS + Coloc, or cS2G + Coloc) and formats the final outputs.
}
