% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtl_data_preparation.R
\name{eqtl_data_preparation}
\alias{eqtl_data_preparation}
\title{Prepare the eQTL data used in the colocalization-based linking}
\usage{
eqtl_data_preparation(
  eqtl_data_folder,
  eqtl_data_tissue = NULL,
  smr_format = TRUE,
  eqtl_data_suffix = NULL,
  save_folder = NULL,
  smr_exe,
  extract_gene_list,
  smr_file_delete = TRUE,
  complete_file_delete = TRUE,
  maf_impute = TRUE,
  maf_impute_file
)
}
\arguments{
\item{maf_impute_file}{}
}
\description{
Yanglab provides a tidy dataset for GTEx v8 cis eQTLs, but this dataset lacks MAF information. Therefore, this function supplies MAF values (sourced from the “eQTL_catalogue”) to complete the GTEx summary data. In addition, this program splits the eQTL data by individual genes.
}
