% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_data_preparation.R
\name{gwas_data_preparation}
\alias{gwas_data_preparation}
\title{Cleaning the GWAS summary data for colocalization}
\usage{
gwas_data_preparation(
  pheno_name,
  SNP_data_create_above,
  gwas_data_folder,
  gwas_data_suffix,
  gwas_allele_input,
  gwas_data_type,
  gwas_sdy_input = NULL,
  gwas_s_input = NULL,
  gwas_beta_input,
  save_folder
)
}
\arguments{
\item{save_folder}{}
}
\description{
This program is designed to convert GWAS summary data intended for subsequent colocalization analyses into a specific tidy format. Given the GWAS phenotype categories and the information available, users must select one of the following input configurations based on their particular circumstances (see details).
}
\details{
For quantitative GWAS phenotypes:

Specify gwas_data_type = "quant", and then choose one of the following scenarios:

1.With phenotype standard deviation (sdY) and variants' beta and standard error:
Please provide the "gwas_sdy_input" parameter, and ensure the GWAS summary data includes the following columns: SNP, beta, SE.
Note: Although the coloc package requires varbeta (i.e., the square of SE), only SE needs to be provided here. The same applies to subsequent cases.

2.With phenotype standard deviation (sdY) but without variants' beta and standard error:
Please provide the "gwas_sdy_input" parameter and ensure the GWAS summary data includes the following columns: SNP, pvalues, MAF.

3.Without phenotype standard deviation (sdY) but with variants' beta and standard error:
Please ensure the GWAS summary data includes the following columns: SNP, beta, SE, N, MAF.

4.Without phenotype standard deviation (sdY) and without variants' beta and standard error:
Please ensure the GWAS summary data includes the following columns: SNP, N, MAF, pvalues.

For binary (Case-Control) GWAS phenotypes:

Specify gwas_data_type = "cc", and provide the "gwas_s_input" parameter (requisite). Then, choose one of the following scenarios:

1.With variants' beta and standard error:
Please ensure nsure the GWAS summary data includes the following columns: SNP, beta, SE.

2.Without variants' beta and standard error:
Please ensure the GWAS summary data includes the following columns: SNP, pvalues, MAF.

It should be noted that if you wish to align the strands between GWAS and eQTL in the future, you need to set "gwas_allele_input" to TRUE and provide additional column in GWAS summary: A1 (effect allele) and A2 (other allele).
}
