% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nominated_gene_extract.R
\name{nominated_gene_extract}
\alias{nominated_gene_extract}
\title{Extract nominated genes for colocalization-based linking}
\usage{
nominated_gene_extract(
  pheno_name,
  cs2g_result_folder,
  pops_result_folder,
  save_folder
)
}
\arguments{
\item{pheno_name}{Character string or vector. Comma-separated or vector of phenotype names.}

\item{cs2g_result_folder}{Path to the folder containing cS2G results.}

\item{pops_result_folder}{Path to the folder containing PoPs results.}

\item{save_folder}{Path where the output files will be saved.}
}
\description{
Use this function to organize the list of potentially relevant genes obtained from locus-based linking (cS2G) and similarity-based linking (PoPs).
Low confidence cS2G genes (< 0.8) and all PoPs genes are nominated for colocalization.
High confidence cS2G genes (>= 0.8) are saved separately as credible genes.
}
