% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pops_result_process.R
\name{pops_result_processing}
\alias{pops_result_processing}
\title{Clean the PoPs result}
\usage{
pops_result_processing(
  pheno_name,
  SNP_data,
  pops_result_folder,
  save_folder,
  pops_result_suffix = NULL,
  annot_file = NULL,
  ID_replace = TRUE,
  replace_list = NULL,
  ID_remove = TRUE,
  remove_list = NULL
)
}
\arguments{
\item{remove_list}{}
}
\description{
Use this function to convert gene names or delete  specific genes (such as non protein coding genes); You can specify the specific list to be converted or deleted by yourself
}
