% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_process.R
\name{result_process}
\alias{result_process}
\title{Process results from mdS2G}
\usage{
result_process(
  coloc_result,
  SNP_data_create_above,
  cs2g_confi_data,
  save_folder
)
}
\arguments{
\item{save_folder}{}
}
\description{
This function is designed to process and summarize the results previously obtained from mdS2G.
}
